requirements.txt
playwright>=1.40
httpx>=0.24
python-dotenv>=1.0
